move filterAssistantResults() function from assistantSaga to Assistant/utils
